+++ NEWS.md
@@ -2,7 +2,9 @@ rticles 0.19
 ---------------------------------------------------------------------
 - Update Copernicus Publications template to version 6.2 from 2021-01-15 (thanks, @RLumSK, #366).
 
-- Add article template `ims_article()` for the *Annals of Applied Statistics* (thanks, @auzaheta, #372)
+- Add article template `pihph_article()` for the *Papers in Historical Phonology* (PiHPh) (thanks, @stefanocoretta, #362).
+
+- Add article template `ims_article()` for *Institute of Mathematical Statistics* Journals, e.g., *Annals of Applied Statistics* (thanks, @auzaheta, #372)
 
 rticles 0.18
 ---------------------------------------------------------------------